Broke out the renderBoard function and dependent functions including renderTile, renderRow, and renderTileImage. #28
